requirements.txt
ftfy
cssutils
gdown
imageio
imageio-ffmpeg
importlib-metadata
ipykernel
ipython
ipython-genutils
json5
jsonpatch
jsonpointer
jsonschema
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
matplotlib
matplotlib-inline
moviepy
notebook
numba
numpy
nvidia-ml-py3
opencv-python
pandas
pathtools
Pillow
pip
plotly
psutil
ptyprocess
pyaml
regex
scikit-image
scikit-learn
scipy
seaborn
subprocess32
svgpathtools
svgwrite
torch
torch-tools
torchfile
torchvision
tqdm
visdom
wandb
webencodings
websocket-client
zipp
ipywidgets
